Move duplicate web utils to common, remove unused notifications.
