+++ podium-gateway/src/main/java/nl/thehyve/podium/web/rest/RequestDetailResource.java
@@ -11,9 +11,9 @@ import com.codahale.metrics.annotation.Timed;
 import io.swagger.annotations.ApiParam;
 import nl.thehyve.podium.common.security.AuthorityConstants;
 import nl.thehyve.podium.common.security.annotations.SecuredByAuthority;
+import nl.thehyve.podium.common.web.rest.util.PaginationUtil;
 import nl.thehyve.podium.domain.RequestDetail;
 import nl.thehyve.podium.service.RequestDetailService;
-import nl.thehyve.podium.web.rest.util.PaginationUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
