+++ podium-gateway/src/main/webapp/app/backoffice/modules/organisation/organisation-delete-dialog.component.html
@@ -16,7 +16,7 @@
     </div>
     <div class="modal-body">
         <pdm-alert-error></pdm-alert-error>
-        <p jhiTranslate="podiumGatewayApp.organisation.delete.question" translateValues="{name: '{{organisation.name}}'}">Are you sure you want to delete this Organisation?</p>
+        <p jhiTranslate="organisation.delete.question" translateValues="{name: '{{organisation.name}}'}">Are you sure you want to delete this Organisation?</p>
     </div>
     <div class="modal-footer">
         <button type="button" class="btn btn-secondary" data-dismiss="modal" (click)="clear()">
