+++ podium-gateway/src/main/webapp/app/blocks/interceptor/http.provider.ts
@@ -14,7 +14,6 @@ import { AuthInterceptor } from './auth.interceptor';
 import { LocalStorageService, SessionStorageService } from 'ng2-webstorage';
 import { AuthExpiredInterceptor } from './auth-expired.interceptor';
 import { ErrorHandlerInterceptor } from './errorhandler.interceptor';
-import { NotificationInterceptor } from './notification.interceptor';
 
 export function interceptableFactory(
     backend: XHRBackend,
@@ -32,7 +31,6 @@ export function interceptableFactory(
             new AuthExpiredInterceptor(injector),
             // Other interceptors can be added here
             new ErrorHandlerInterceptor(eventManager),
-            new NotificationInterceptor()
         ]
     );
 }
