--- podium-gateway/src/main/webapp/app/blocks/interceptor/notification.interceptor.ts
@@ -1,42 +1,0 @@
-/*
- * Copyright (c) 2017. The Hyve and respective contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- *
- * See the file LICENSE in the root of this repository.
- *
- */
-import { JhiHttpInterceptor } from 'ng-jhipster';
-import { RequestOptionsArgs, Response } from '@angular/http';
-import { Observable } from 'rxjs/Observable';
-
-export class NotificationInterceptor extends JhiHttpInterceptor {
-
-    constructor() {
-        super();
-    }
-
-    requestIntercept(options?: RequestOptionsArgs): RequestOptionsArgs {
-        return options;
-    }
-
-    responseIntercept(observable: Observable<Response>): Observable<Response> {
-        return <Observable<Response>> observable.catch((error) => {
-            let arr = Array.from(error.headers._headers);
-            let headers = [];
-            let i;
-            for (i = 0; i < arr.length; i++) {
-                if (arr[i][0].endsWith('app-alert') || arr[i][0].endsWith('app-params')) {
-                    headers.push(arr[i][0]);
-                }
-            }
-            headers.sort();
-            let alertKey = headers.length >= 1 ? error.headers.get(headers[0]) : null;
-            if (typeof alertKey === 'string') {
-                // AlertService.success(alertKey, { param: response.headers(headers[1])});
-            }
-            return Observable.throw(error);
-        });
-    }
-}
