+++ podium-gateway/src/main/webapp/app/shared/alert/alert-error.component.ts
@@ -11,6 +11,7 @@ import { Component, OnDestroy } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import { JhiEventManager, JhiAlertService } from 'ng-jhipster';
 import { Subscription } from 'rxjs/Rx';
+import { FieldError } from './field-error';
 
 @Component({
     selector: 'pdm-alert-error',
@@ -32,7 +33,6 @@ export class PdmAlertErrorComponent implements OnDestroy {
         this.alerts = [];
 
         this.cleanHttpErrorListener = eventManager.subscribe('podiumGatewayApp.httpError', (response) => {
-            let i;
             let httpResponse = response.content;
             switch (httpResponse.status) {
                 // connection refused, server not reachable
@@ -41,30 +41,15 @@ export class PdmAlertErrorComponent implements OnDestroy {
                     break;
 
                 case 400:
-                    let arr = Array.from(httpResponse.headers._headers);
-                    let headers = [];
-                    for (i = 0; i < arr.length; i++) {
-                        if (arr[i][0].endsWith('app-error') || arr[i][0].endsWith('app-params')) {
-                            headers.push(arr[i][0]);
-                        }
-                    }
-                    headers.sort();
-                    let errorHeader = httpResponse.headers.get(headers[0]);
-                    let entityKey = httpResponse.headers.get(headers[1]);
-                    if (errorHeader) {
-                        let entityName = translateService.instant('global.menu.entities.' + entityKey);
-                        this.addErrorAlert(errorHeader, errorHeader, {entityName: entityName});
-                    } else if (httpResponse.text() !== '' && httpResponse.json() && httpResponse.json().fieldErrors) {
-                        let fieldErrors = httpResponse.json().fieldErrors;
-                        for (i = 0; i < fieldErrors.length; i++) {
-                            let fieldError = fieldErrors[i];
+                    if (httpResponse.text() !== '' && httpResponse.json() && httpResponse.json().fieldErrors) {
+                        let fieldErrors = httpResponse.json().fieldErrors as FieldError[];
+                        fieldErrors.forEach((fieldError) => {
                             // convert 'something[14].other[4].id' to 'something[].other[].id' so translations can be written to it
                             let convertedField = fieldError.field.replace(/\[\d*\]/g, '[]');
-                            let fieldName = translateService.instant('podiumGatewayApp.' +
-                                fieldError.objectName + '.' + convertedField);
+                            let fieldName = translateService.instant(fieldError.objectName + '.' + convertedField);
                             this.addErrorAlert(
                                 'Field ' + fieldName + ' cannot be empty', 'error.' + fieldError.message, {fieldName: fieldName});
-                        }
+                        });
                     } else if (httpResponse.text() !== '' && httpResponse.json() && httpResponse.json().message) {
                         this.addErrorAlert(httpResponse.json().message, httpResponse.json().message, httpResponse.json());
                     } else {
