+++ podium-gateway/src/main/webapp/app/shared/alert/field-error.ts
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2018. The Hyve and respective contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ * See the file LICENSE in the root of this repository.
+ *
+ */
+export class FieldError {
+    objectName: string;
+    field: string;
+    message: string;
+}
