+++ podium-gateway/src/main/webapp/app/shared/role/role-assign/role-assign.component.html
@@ -8,7 +8,7 @@
   ~
   -->
 
-<h4><span jhiTranslate="podiumGatewayApp.role.home.title">Roles</span></h4>
+<h4><span jhiTranslate="role.home.title">Roles</span></h4>
 
 <div class="container-fluid" *ngIf="organisationRoles">
     <div class="row">
@@ -22,9 +22,9 @@
                 <table class="table">
                     <thead>
                         <tr>
-                            <th jhiTranslate="podiumGatewayApp.role.users">User</th>
-                            <th jhiTranslate="podiumGatewayApp.role.role">Role</th>
-                            <th jhiTranslate="podiumGatewayApp.role.options">Options</th>
+                            <th jhiTranslate="role.users">User</th>
+                            <th jhiTranslate="role.role">Role</th>
+                            <th jhiTranslate="role.options">Options</th>
                         </tr>
                     </thead>
                     <tbody>
