+++ podium-gateway/src/main/webapp/i18n/en/role.json
@@ -1,35 +1,29 @@
 {
-    "podiumGatewayApp": {
-        "role" : {
-            "home": {
-                "title": "Roles",
-                "createLabel": "Create a new Role",
-                "createOrEditLabel": "Create or edit a Role",
-                "search": "Search for Role"
-            },
-            "delete": {
-                "question": "Are you sure you want to delete Role {{ id }}?"
-            },
-            "detail": {
-                "title": "Role"
-            },
-            "users": "Users",
-            "user": "User",
-            "role": "Role",
-            "organisation": "Organisation",
-            "authority": "Authority",
-            "options": "Options"
+    "role" : {
+        "home": {
+            "title": "Roles",
+            "createLabel": "Create a new Role",
+            "createOrEditLabel": "Create or edit a Role",
+            "search": "Search for Role"
         },
-        "roleAssign": {
-            "saved": "Successfully saved",
-            "deleted": "Successfully deleted"
-        }
+        "delete": {
+            "question": "Are you sure you want to delete Role {{ id }}?"
+        },
+        "detail": {
+            "title": "Role"
+        },
+        "users": "Users",
+        "user": "User",
+        "role": "Role",
+        "organisation": "Organisation",
+        "authority": "Authority",
+        "options": "Options",
+        "created": "A new Role is created with identifier {{ param }}",
+        "updated": "A Role is updated with identifier {{ param }}",
+        "deleted": "A Role is deleted with identifier {{ param }}"
     },
-    "podiumuaaApp": {
-        "role" : {
-            "created": "A new Role is created with identifier {{ param }}",
-            "updated": "A Role is updated with identifier {{ param }}",
-            "deleted": "A Role is deleted with identifier {{ param }}"
-        }
+    "roleAssign": {
+        "saved": "Successfully saved",
+        "deleted": "Successfully deleted"
     }
 }
