+++ podium-uaa/src/main/java/nl/thehyve/podium/web/rest/AuditResource.java
@@ -11,9 +11,7 @@ import io.swagger.annotations.ApiParam;
 import nl.thehyve.podium.common.security.AuthorityConstants;
 import nl.thehyve.podium.common.security.annotations.SecuredByAuthority;
 import nl.thehyve.podium.service.AuditEventService;
-import nl.thehyve.podium.web.rest.util.PaginationUtil;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import nl.thehyve.podium.common.web.rest.util.PaginationUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.actuate.audit.AuditEvent;
 import org.springframework.data.domain.Page;
@@ -29,7 +27,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.net.URISyntaxException;
-import java.time.LocalDate;
 import java.util.Date;
 import java.util.List;
