+++ podium-uaa/src/main/java/nl/thehyve/podium/web/rest/ElasticsearchIndexResource.java
@@ -18,8 +18,8 @@ import com.codahale.metrics.annotation.Timed;
 import nl.thehyve.podium.common.security.AuthorityConstants;
 import nl.thehyve.podium.common.security.annotations.SecuredByAuthority;
 import nl.thehyve.podium.common.service.SecurityService;
+import nl.thehyve.podium.common.web.rest.util.HeaderUtil;
 import nl.thehyve.podium.service.ElasticsearchIndexService;
-import nl.thehyve.podium.web.rest.util.HeaderUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
