+++ podium-uaa/src/main/java/nl/thehyve/podium/web/rest/RoleResource.java
@@ -17,15 +17,14 @@ import nl.thehyve.podium.common.security.annotations.OrganisationUuidParameter;
 import nl.thehyve.podium.common.security.annotations.SecuredByAuthority;
 import nl.thehyve.podium.common.security.annotations.SecuredByOrganisation;
 import nl.thehyve.podium.common.service.dto.RoleRepresentation;
+import nl.thehyve.podium.common.web.rest.util.HeaderUtil;
 import nl.thehyve.podium.domain.Role;
 import nl.thehyve.podium.domain.User;
 import nl.thehyve.podium.service.OrganisationService;
 import nl.thehyve.podium.service.RoleService;
 import nl.thehyve.podium.service.UserService;
-import nl.thehyve.podium.common.service.dto.RoleRepresentation;
 import nl.thehyve.podium.service.mapper.RoleMapper;
-import nl.thehyve.podium.web.rest.util.HeaderUtil;
-import nl.thehyve.podium.web.rest.util.PaginationUtil;
+import nl.thehyve.podium.common.web.rest.util.PaginationUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
