+++ podium-uaa/src/main/java/nl/thehyve/podium/web/rest/UserResource.java
@@ -18,6 +18,7 @@ import nl.thehyve.podium.common.security.annotations.SecuredByAuthority;
 import nl.thehyve.podium.common.security.annotations.SecuredByOrganisation;
 import nl.thehyve.podium.common.service.SecurityService;
 import nl.thehyve.podium.common.service.dto.UserRepresentation;
+import nl.thehyve.podium.common.web.rest.util.HeaderUtil;
 import nl.thehyve.podium.domain.User;
 import nl.thehyve.podium.exceptions.EmailAddressAlreadyInUse;
 import nl.thehyve.podium.exceptions.LoginAlreadyInUse;
@@ -27,8 +28,7 @@ import nl.thehyve.podium.service.MailService;
 import nl.thehyve.podium.service.UserService;
 import nl.thehyve.podium.service.mapper.UserMapper;
 import nl.thehyve.podium.web.rest.dto.ManagedUserRepresentation;
-import nl.thehyve.podium.web.rest.util.HeaderUtil;
-import nl.thehyve.podium.web.rest.util.PaginationUtil;
+import nl.thehyve.podium.common.web.rest.util.PaginationUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
