--- podium-uaa/src/main/java/nl/thehyve/podium/web/rest/util/HeaderUtil.java
@@ -1,50 +1,0 @@
-/*
- * Copyright (c) 2017  The Hyve and respective contributors.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * See the file LICENSE in the root of this repository.
- */
-
-package nl.thehyve.podium.web.rest.util;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpHeaders;
-
-/**
- * Utility class for HTTP headers creation.
- */
-public final class HeaderUtil {
-
-    private static final Logger log = LoggerFactory.getLogger(HeaderUtil.class);
-
-    private HeaderUtil() {
-    }
-
-    public static HttpHeaders createAlert(String message, String param) {
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("X-podiumUaaApp-alert", message);
-        headers.add("X-podiumUaaApp-params", param);
-        return headers;
-    }
-
-    public static HttpHeaders createEntityCreationAlert(String entityName, String param) {
-        return createAlert("podiumUaaApp." + entityName + ".created", param);
-    }
-
-    public static HttpHeaders createEntityUpdateAlert(String entityName, String param) {
-        return createAlert("podiumUaaApp." + entityName + ".updated", param);
-    }
-
-    public static HttpHeaders createEntityDeletionAlert(String entityName, String param) {
-        return createAlert("podiumUaaApp." + entityName + ".deleted", param);
-    }
-
-    public static HttpHeaders createFailureAlert(String entityName, String errorKey, String defaultMessage) {
-        log.error("Entity creation failed, {}", defaultMessage);
-        HttpHeaders headers = new HttpHeaders();
-        headers.add("X-podiumUaaApp-error", "error." + errorKey);
-        headers.add("X-podiumUaaApp-params", entityName);
-        return headers;
-    }
-}
